requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
Deprecated==1.2.14
exceptiongroup==1.2.0
fonttools==4.49.0
fpdf2==2.7.7
h11==0.14.0
idna==3.6
lxml==5.1.0
outcome==1.3.0.post0
packaging==23.2
pillow==10.2.0
PyMuPDF==1.23.22
PyMuPDFb==1.23.22
pypdf==4.0.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
selenium==4.17.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.0
webdriver-manager==4.0.1
wrapt==1.16.0
wsproto==1.2.0
